WebMIDIAPITest/webmidiapi_01_request.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: requestMIDIAccess()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#obtaining-access-to-midi-devices">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="flags" content="WebMIDIAPI.js">
    <meta name="assert" content="Return true when test success in requestMIDIAccess().">
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
    <script src="Uint8Array.js"></script>
    <script src="WebMIDIAPI.js"></script>
    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    </style>
  </head>
  <body>
    <p>Success when requestMIDIAccess() exists and successfully connected.</p>
    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <!--script type="text/javascript" src="webmidiapi_01_request.js"></script-->
    <script type="text/javascript">
    var MIDI;

    var test_requestMIDIAccess = async_test("requestMIDIAccess() - W3C Spec: Web MIDI API");

    function testWMA() {
        navigator.requestMIDIAccess({sysex:false}).then( successCallback, errorCallback );
        function successCallback(access) {
            MIDI=access;
            test_requestMIDIAccess.step(function() {
                assert_true(true);
            });
            test_requestMIDIAccess.done();
        }
        function errorCallback() {
            test_requestMIDIAccess.step(function() {
                assert_false(true);
                test_requestMIDIAccess.done();
            });
        }
    }

    testWMA();
    </script>


  </body>
</html>
